[Bronze III] Title: ZOAC 4, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/23971. ZOAC 4/ZOAC 4.cc
@@ -6,6 +6,6 @@ double h, w, n, m;
 
 int main() {
     cin >> h >> w >> n >> m;
-    cout << (long long)ceil(h/++n) * (long long)ceil(w/++m);
+    cout << (int)(ceil(h/++n) * ceil(w/++m));
     return 0;
 }
